refactor!: remove deprecated GeneratedVaadinComboBox
